move generateHash out of directus_users
